wip: split canonical policy infrastructure
